2025-spring/exp04-easy-rag-system/requirements.txt
fastapi
uvicorn
jinja2
faiss-cpu
sentence-transformers
transformers
torch
accelerate
beautifulsoup4
pydantic
pandas
matplotlib
seaborn
